MM-67322 Add i18n-extract support for shared package and move key constants
