+++ webapp/channels/src/components/with_tooltip/tooltip_shortcut.test.tsx
@@ -2,13 +2,12 @@
 // See LICENSE.txt for license information.
 
 import React from 'react';
-import {defineMessage} from 'react-intl';
 
 import * as userAgentUtils from '@mattermost/shared/utils/user_agent';
 
 import {renderWithContext, screen} from 'tests/react_testing_utils';
 
-import TooltipShortcut from './tooltip_shortcut';
+import TooltipShortcut, {ShortcutKeys} from './tooltip_shortcut';
 
 jest.mock('@mattermost/shared/utils/user_agent', () => ({
     isMac: jest.fn(),
@@ -59,10 +58,7 @@ describe('TooltipShortcut', () => {
     test('show shortcut with message descriptor', () => {
         const shortcut = {
             default: [
-                defineMessage({
-                    id: 'shortcuts.generic.enter',
-                    defaultMessage: 'Enter',
-                }),
+                ShortcutKeys.enter,
             ],
         };
